--- src/app/[problemId]/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/containers/[problemId]/GameScreen';
