+++ src/app/problems/[problemId]/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/containers/problems/[problemId]/GameScreen';
