--- src/containers/Home.tsx
@@ -1,7 +1,0 @@
-export const Home = () => {
-	return (
-		<div>
-			<h1>Hello World</h1>
-		</div>
-	);
-};
